+++ integrador/src/main/java/com/backend/integrador/entity/Odontologo.java
@@ -2,12 +2,12 @@ package com.backend.integrador.entity;
 
 public class Odontologo {
 
-    private int id;
+    private Long id;
     private int numeroMatricula;
     private String nombre;
     private String apellido;
 
-    public Odontologo(int id, int numeroMatricula, String nombre, String apellido) {
+    public Odontologo(Long id, int numeroMatricula, String nombre, String apellido) {
         this.id = id;
         this.numeroMatricula = numeroMatricula;
         this.nombre = nombre;
@@ -20,11 +20,11 @@ public class Odontologo {
         this.apellido = apellido;
     }
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
